+++ src/mc_control/samples/LIPMStabilizer/src/mc_lipm_stabilizer.h
@@ -12,15 +12,8 @@ struct MC_CONTROL_DLLAPI LIPMStabilizerController : public mc_control::fsm::Cont
 {
   LIPMStabilizerController(mc_rbdyn::RobotModulePtr rm, double dt, const mc_rtc::Configuration & config);
 
-  bool run() override;
-
-  void reset(const mc_control::ControllerResetData & reset_data) override;
-
   void supported_robots(std::vector<std::string> & out) const override
   {
     out = {"jvrc1", "hrp2_drc", "hrp4", "hrp5_p"};
   }
-
-private:
-  mc_rtc::Configuration config_;
 };