+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 ### Fixes
 - Fix and upgrade benchmarks compilation
+- Correctly teardown the last FSM state
 
 ## [1.3.0] - 2020-05-07